- DB Versioning handled - onResponseReceived interface changed - survey for multi-user done

+++ app/src/main/java/com/oneflow/analytics/controller/EventController.java
@@ -39,7 +39,7 @@ public class EventController implements MyResponseHandler {
     }
 
     @Override
-    public void onResponseReceived(Constants.ApiHitType hitType, Object obj, int reserve) {
+    public void onResponseReceived(Constants.ApiHitType hitType, Object obj, Long reserve) {
 
         switch(hitType){
             case insertEventsInDB:
