+++ app/src/main/java/com/oneflow/analytics/utils/MyResponseHandler.java
@@ -1,5 +1,5 @@
 package com.oneflow.analytics.utils;
 
 public interface MyResponseHandler {
-    public void onResponseReceived(Constants.ApiHitType hitType, Object obj, int reserve);
+    public void onResponseReceived(Constants.ApiHitType hitType, Object obj, Long reserve);
 }
